fix: staging pulled and merged

+++ pkg/jdbc/jdbc.go
@@ -35,7 +35,7 @@ func MinMaxQuery(stream types.StreamInterface, column string) string {
 //	  ORDER BY id, created_at
 //	  LIMIT 1000
 //	) AS subquery
-func NextChunkEndQuery(stream types.StreamInterface, columns []string, chunkSize int, parsedFilter string) string {
+func NextChunkEndQuery(stream types.StreamInterface, columns []string, chunkSize int64, parsedFilter string) string {
 	var query strings.Builder
 	// SELECT with quoted and concatenated values
 	fmt.Fprintf(&query, "SELECT MAX(key_str) FROM (SELECT CONCAT_WS(',', %s) AS key_str FROM `%s`.`%s`",
